@@ -32,7 +32,6 @@ dependencies:
   cloud_firestore_web: 3.12.3
   collection: 1.18.0
   dropdown_button2: 2.3.9
-  easy_debounce: 2.0.1
   file_picker: 8.0.3
   firebase_auth: 4.19.5
   firebase_auth_platform_interface: 7.2.6
@@ -49,7 +48,7 @@ dependencies:
   flutter_animate: 4.5.0
   flutter_cache_manager: 3.3.1
   flutter_plugin_android_lifecycle: 2.0.20
-  font_awesome_flutter: 10.6.0
+  font_awesome_flutter: 10.7.0
   from_css_color: 2.0.0
   go_router: 12.1.3
   google_fonts: 6.1.0
@@ -87,9 +86,9 @@ dependencies:
   sqflite: 2.2.6
   stream_transform: 2.1.0
   timeago: 3.6.1
-  url_launcher: 6.2.5
-  url_launcher_android: 6.3.0
-  url_launcher_ios: 6.2.5
+  url_launcher: 6.3.0
+  url_launcher_android: 6.3.4
+  url_launcher_ios: 6.3.1
   url_launcher_platform_interface: 2.3.2
   video_player: 2.8.6
   video_player_android: 2.4.16
@@ -133,6 +132,7 @@ flutter:
     - assets/lottie_animations/
     - assets/rive_animations/
     - assets/pdfs/
+
 
   
   # An image asset can refer to one or more resolution-specific "variants", see
