@@ -1,6 +1,10 @@
+import 'dart:async';
 
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
 
+import '/backend/backend.dart';
+import 'package:stream_transform/stream_transform.dart';
 import 'firebase_auth_manager.dart';
 
 export 'firebase_auth_manager.dart';
@@ -8,15 +12,19 @@ export 'firebase_auth_manager.dart';
 final _authManager = FirebaseAuthManager();
 FirebaseAuthManager get authManager => _authManager;
 
-String get currentUserEmail => currentUser?.email ?? '';
+String get currentUserEmail =>
+    currentUserDocument?.email ?? currentUser?.email ?? '';
 
 String get currentUserUid => currentUser?.uid ?? '';
 
-String get currentUserDisplayName => currentUser?.displayName ?? '';
+String get currentUserDisplayName =>
+    currentUserDocument?.displayName ?? currentUser?.displayName ?? '';
 
-String get currentUserPhoto => currentUser?.photoUrl ?? '';
+String get currentUserPhoto =>
+    currentUserDocument?.photoUrl ?? currentUser?.photoUrl ?? '';
 
-String get currentPhoneNumber => currentUser?.phoneNumber ?? '';
+String get currentPhoneNumber =>
+    currentUserDocument?.phoneNumber ?? currentUser?.phoneNumber ?? '';
 
 String get currentJwtToken => _currentJwtToken ?? '';
 
@@ -29,3 +37,34 @@ final jwtTokenStream = FirebaseAuth.instance
     .idTokenChanges()
     .map((user) async => _currentJwtToken = await user?.getIdToken())
     .asBroadcastStream();
+
+DocumentReference? get currentUserReference =>
+    loggedIn ? UsersRecord.collection.doc(currentUser!.uid) : null;
+
+UsersRecord? currentUserDocument;
+final authenticatedUserStream = FirebaseAuth.instance
+    .authStateChanges()
+    .map<String>((user) => user?.uid ?? '')
+    .switchMap(
+      (uid) => uid.isEmpty
+          ? Stream.value(null)
+          : UsersRecord.getDocument(UsersRecord.collection.doc(uid))
+              .handleError((_) {}),
+    )
+    .map((user) {
+  currentUserDocument = user;
+
+  return currentUserDocument;
+}).asBroadcastStream();
+
+class AuthUserStreamWidget extends StatelessWidget {
+  const AuthUserStreamWidget({super.key, required this.builder});
+
+  final WidgetBuilder builder;
+
+  @override
+  Widget build(BuildContext context) => StreamBuilder(
+        stream: authenticatedUserStream,
+        builder: (context, _) => builder(context),
+      );
+}
