@@ -132,8 +132,8 @@ class _PostConfirmationDialogWidgetState
                         child: Align(
                           alignment: const AlignmentDirectional(0.0, 1.0),
                           child: FFButtonWidget(
-                            onPressed: () {
-                              print('Button pressed ...');
+                            onPressed: () async {
+                              context.goNamed('orgHomepage');
                             },
                             text: 'OK!',
                             options: FFButtonOptions(
