@@ -1,3 +1,4 @@
+import '/components/bottom_navigation_bar_org_widget.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import 'org_profile_copy_copy_widget.dart' show OrgProfileCopyCopyWidget;
 import 'package:flutter/material.dart';
@@ -6,14 +7,21 @@ class OrgProfileCopyCopyModel
     extends FlutterFlowModel<OrgProfileCopyCopyWidget> {
   ///  State fields for stateful widgets in this page.
 
+  // Model for BottomNavigationBarOrg component.
+  late BottomNavigationBarOrgModel bottomNavigationBarOrgModel;
   bool isDataUploading = false;
   FFUploadedFile uploadedLocalFile =
       FFUploadedFile(bytes: Uint8List.fromList([]));
   String uploadedFileUrl = '';
 
   @override
-  void initState(BuildContext context) {}
+  void initState(BuildContext context) {
+    bottomNavigationBarOrgModel =
+        createModel(context, () => BottomNavigationBarOrgModel());
+  }
 
   @override
-  void dispose() {}
+  void dispose() {
+    bottomNavigationBarOrgModel.dispose();
+  }
 }
