@@ -1,4 +1,5 @@
 import '/backend/backend.dart';
+import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import '/flutter_flow/flutter_flow_widgets.dart';
@@ -46,6 +47,36 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
   Widget build(BuildContext context) {
     return Scaffold(
       key: scaffoldKey,
+      appBar: AppBar(
+        backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
+        automaticallyImplyLeading: false,
+        leading: FlutterFlowIconButton(
+          borderColor: Colors.transparent,
+          borderRadius: 30.0,
+          borderWidth: 1.0,
+          buttonSize: 60.0,
+          icon: Icon(
+            Icons.arrow_back_rounded,
+            color: FlutterFlowTheme.of(context).primaryText,
+            size: 30.0,
+          ),
+          onPressed: () async {
+            context.pop();
+          },
+        ),
+        title: Text(
+          'Page Title',
+          style: FlutterFlowTheme.of(context).headlineMedium.override(
+                fontFamily: 'Inter Tight',
+                color: FlutterFlowTheme.of(context).primaryText,
+                fontSize: 22.0,
+                letterSpacing: 0.0,
+              ),
+        ),
+        actions: const [],
+        centerTitle: true,
+        elevation: 0.0,
+      ),
       body: StreamBuilder<ChallengesRecord>(
         stream: ChallengesRecord.getDocument(widget.chalRef!),
         builder: (context, snapshot) {
@@ -76,20 +107,6 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
                     mainAxisSize: MainAxisSize.min,
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      InkWell(
-                        splashColor: Colors.transparent,
-                        focusColor: Colors.transparent,
-                        hoverColor: Colors.transparent,
-                        highlightColor: Colors.transparent,
-                        onTap: () async {
-                          context.pop();
-                        },
-                        child: Icon(
-                          Icons.arrow_circle_left_sharp,
-                          color: FlutterFlowTheme.of(context).primaryText,
-                          size: 32.0,
-                        ),
-                      ),
                       ClipRRect(
                         borderRadius: BorderRadius.circular(18.0),
                         child: CachedNetworkImage(
@@ -194,9 +211,7 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
                                                       .override(
                                                         fontFamily: 'Outfit',
                                                         color:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .primary,
+                                                            const Color(0xFF0043CE),
                                                         fontSize: 14.0,
                                                         letterSpacing: 0.0,
                                                         fontWeight:
@@ -206,7 +221,10 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
                                           ],
                                         ),
                                         Text(
-                                          'Moments ago',
+                                          dateTimeFormat(
+                                              "yMMMd",
+                                              columnChallengesRecord
+                                                  .createdTime!),
                                           style: FlutterFlowTheme.of(context)
                                               .labelMedium
                                               .override(
@@ -298,9 +316,7 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
                                                       .override(
                                                         fontFamily: 'Outfit',
                                                         color:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .primary,
+                                                            const Color(0xFF0043CE),
                                                         fontSize: 14.0,
                                                         letterSpacing: 0.0,
                                                         fontWeight:
@@ -310,7 +326,10 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
                                           ],
                                         ),
                                         Text(
-                                          'Moments ago',
+                                          dateTimeFormat(
+                                              "d/M h:mm a",
+                                              columnChallengesRecord
+                                                  .createdTime!),
                                           style: FlutterFlowTheme.of(context)
                                               .labelMedium
                                               .override(
@@ -355,8 +374,7 @@ class _SolDetailsWidgetState extends State<SolDetailsWidget> {
                                           iconPadding:
                                               const EdgeInsetsDirectional.fromSTEB(
                                                   0.0, 0.0, 0.0, 0.0),
-                                          color: FlutterFlowTheme.of(context)
-                                              .primary,
+                                          color: const Color(0xFF0043CE),
                                           textStyle:
                                               FlutterFlowTheme.of(context)
                                                   .titleSmall
